Effort-Weighted Log Gaussian Cox Process SDM

Runs the effort-weighted Log Gaussian Cox Process species distribution model

Usage

ewlgcp(
  formula,
  dmesh,
  effort = TRUE,
  adjust = TRUE,
  buffer = TRUE,
  orthogonal = TRUE,
  prior.beta = NULL,
  prior.range = c(50, 0.1),
  prior.sigma = c(1, 0.1),
  smooth = 3/2,
  ...
)

Arguments

formula

A formula of the form y ~ x1 + x2 + ....

dmesh

A list with elements produced by the different dmesh_ functions.

effort

Logical. Whether to adjust the model for effort. Default TRUE.

adjust

Logical. Whether to adjust the effort for being species specific. Default TRUE.

buffer

Logical. Whether to add the effort buffer to help reducing prediction outside of the species range. Default TRUE.

orthogonal

Logical. Whether to make the spatial field orthogonal to the predictors? Default TRUE.

prior.beta

Normal priors for the betas of the fixed effects coefficients as required by INLA. Default is list(prec=list(default=1/(1)^2,Intercept=1/(20)^2),mean=list(default=0,Intercept=0)) which means a prior with mean = 0 and sd = 1 for all coefficients and a prior with mean = 0 and sd = 20 for the model intercept. See ?control.fixed.

prior.range

Penalized complexity prior for the range of the spatial field. A vector of length two giving the probability that the range is inferior to a given value. The default is prior.range = c(50, 0.01) which represents a 1% chance that the range is inferior to 50 (in the units of the crs used).

prior.sigma

Penalized complexity prior for the standard deviation (sd) of the spatial field. A vector of length two giving the probability that the sd is superior to a given value. The default is prior.sigma = c(1, 0.01) which represents a 1% chance that the range is superior to 1.

smooth

x

...

Further arguments to pass to inla

Value

A model of class INLA.

Details

none

References

Simpson, D. Illian, J. B., Lindgren, F. S, S. H. and Rue, H. 2016. Going off grid: computationally efficient inference for log-Gaussian Cox processes. Biometrika, 103(1): 49-70 https://doi.org/10.1093/biomet/asv064

Fuglstad, G.-A., Simpson, D., Lindgren, F. & Rue, H. 2019 Constructing Priors that Penalize the Complexity of Gaussian Random Fields. Journal of the American Statistical Association, 114(525): 445-452 https://doi.org/10.1080/01621459.2017.1415907